requirements.txt
Pillow==10.1.0
PyYAML==6.0.1
requests==2.25.1
